[Decomp][Chat] Cohesion split: extract message validation to ChatMessageValidation.cpp
